requirements.txt
Django
psycopg2
Pillow
python-a2s
django-background-tasks